+++ actix-web-lab/src/respond.rs
@@ -2,5 +2,8 @@
 
 pub use crate::{csv::Csv, display_stream::DisplayStream, html::Html, ndjson::NdJson};
 
+#[cfg(feature = "cbor")]
+pub use crate::cbor::Cbor;
+
 #[cfg(feature = "msgpack")]
 pub use crate::msgpack::{MessagePack, MessagePackNamed};