+++ actix-web-lab/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `middleware::map_response()` for mapping responses with an async function.
 - Add `middleware::map_response_body()` for mapping response bodies with an async function.
 - Add `respond::{MessagePack,MessagePackNamed}` responders.
+- Add `respond::Cbor` responder.
 
 ## 0.18.8 - 2022-12-10